Use views toggle instead of create & destroy

+++ app/controllers/views_controller.rb
@@ -1,12 +1,14 @@
 class ViewsController < ApplicationController
-  def create
-    View.create(view_params.merge(user: current_user))
-    redirect_to :back
-  end
 
-  def destroy
-    View.where(user: current_user, episode_id: params[:id]).destroy_all
-    redirect_to :back
+  def toggle
+    created = false
+    if current_user.seen_episode_id?(view_params[:episode_id])
+      View.where(user: current_user, episode_id: params[:id]).destroy_all
+    else
+      View.create(view_params.merge(user: current_user))
+      created = true
+    end
+    render json: { episode_id: view_params[:episode_id], created: created }, status: 200
   end
 
   def set_all
